pouzite mappera v controllery a jeho test

+++ src/main/java/com/example/demo/model/PersonJpa.java
@@ -26,14 +26,14 @@ public class PersonJpa {
 
     private Priority priority;
 
-    private BigDecimal number;
+    private Long number;
 
     public PersonJpa() {};
 
     public PersonJpa(String firstName, String lastName, BigDecimal number) {
         this.firstName = firstName;
         this.lastName = lastName;
-        this.number = number;
+        this.number = number.longValue();
     }
 
     public PersonJpa(String firstName, String lastName) {
